# api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: Warcraft Addon Download Tracker
paths:
  /addons:
    x-swagger-router-controller: addons
    get:
      summary: Returns the list of all addons that have download counts.
      description: Get the list of addons
      operationId: getAddons
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/addons"
  "/addons/{addonName}":
    x-swagger-router-controller: addons
    get:
      summary: Returns a specific addon
      description: Gets a specific addon
      operationId: getAddon
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/addon"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Creates a new addon
      description: Creates an addon
      operationId: createAddon
      parameters:
        - name: api-token
          in: header
          description: The API token for authorization
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/addon_post"
        description: The addon URLs
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/addon"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    parameters:
      - $ref: "#/components/parameters/addonNameParam"
  "/addons/{addonName}/downloads":
    x-swagger-router-controller: addons
    get:
      summary: Returns download information for {addonName}
      description: Get the list of addons
      operationId: getDownloadsForAddon
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/addonDownloads"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Adds a new download count for {addonName}
      operationId: addDownloadsForAddon
      parameters:
        - name: api-token
          in: header
          description: The API token for authorization
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/addon_downloads_post"
        description: The new download count and timestamp for {addonName}
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/addon_downloads_post_response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    parameters:
      - $ref: "#/components/parameters/addonNameParam"
servers:
  - url: http://localhost:10010/
  - url: https://localhost:10010/
components:
  parameters:
    addonNameParam:
      name: addonName
      in: path
      required: true
      description: The name of an addon to retrieve data from
      schema:
        type: string
  schemas:
    addon:
      type: object
      required:
        - addonName
      properties:
        addonName:
          type: string
        curseForgeUrl:
          type: string
        wowInterfaceUrl:
          type: string
    addon_post:
      type: object
      required:
        - curseForgeUrl
        - wowInterfaceUrl
      properties:
        curseForgeUrl:
          type: string
        wowInterfaceUrl:
          type: string
    addon_downloads_post:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          format: int32
          minimum: 0
    addons:
      properties:
        message:
          type: string
    addonDownloads:
      required:
        - addonName
      type: object
      properties:
        addonName:
          type: string
        downloads:
          type: array
          items:
            $ref: "#/components/schemas/download_count"
    download_count:
      type: object
      required:
        - count
        - timestamp
      properties:
        count:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
    addon_downloads_post_response:
      required:
        - addonName
        - count
        - timestamp
      properties:
        addonName:
          type: string
        count:
          type: integer
          format: int32
          minimum: 0
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      properties:
        message:
          type: string